// enum/enum.ts
export enum Permissions {
  DashboardView = "dashboard.view",

  BoardView = "board.view",
  BoardTaskCreate = "board.task.create",
  BoardTaskEdit = "board.task.edit",
  BoardTaskStatusEdit = "board.task.status.edit",
  BoardTaskDelete = "board.task.delete",
  BoardTaskBoardCreate = "board.task.board.create",
  BoardTaskBoardEdit = "board.task.board.edit",
  BoardTaskBoardDelete = "board.task.board.delete",

  MemberView = "member.view",
}